Conflicts:
	.classpath
	build-clean.xml
	src/freenet/io/comm/PeerContext.java
	src/freenet/io/xfer/BlockTransmitter.java
	src/freenet/io/xfer/BulkTransmitter.java
	src/freenet/io/xfer/OldPacketThrottle.java
	src/freenet/io/xfer/ThrottleDeprecatedException.java
	src/freenet/node/PacketTracker.java
	src/freenet/node/PeerNode.java
	src/freenet/node/Version.java
	src/freenet/node/updater/NodeUpdateManager.java
	src/freenet/node/updater/UpdateDeployContext.java
	src/freenet/support/FileLoggerHook.java
	test/freenet/node/NullBasePeerNode.java

+++ src/freenet/node/BaseSender.java
@@ -9,8 +9,6 @@ import freenet.io.comm.DisconnectedException;
 import freenet.io.comm.Message;
 import freenet.io.comm.MessageFilter;
 import freenet.io.comm.NotConnectedException;
-import freenet.io.comm.PeerContext;
-import freenet.io.comm.SlowAsyncMessageFilterCallback;
 import freenet.keys.Key;
 import freenet.keys.NodeCHK;
 import freenet.keys.NodeSSK;
@@ -19,9 +17,8 @@ import freenet.node.PeerNode.SlotWaiter;
 import freenet.node.PeerNode.SlotWaiterFailedException;
 import freenet.support.LogThresholdCallback;
 import freenet.support.Logger;
-import freenet.support.TimeUtil;
 import freenet.support.Logger.LogLevel;
-import freenet.support.io.NativeThread;
+import freenet.support.TimeUtil;
 
 /** Base class for request and insert senders.
  * Mostly concerned with what happens *before and up to* we get the Accepted.
@@ -83,7 +80,7 @@ public abstract class BaseSender implements ByteCounter {
 	
 	private short hopsForTime(long time) {
 		double timeout = realTimeFlag ? SEARCH_TIMEOUT_REALTIME : SEARCH_TIMEOUT_BULK;
-		double timePerHop = timeout / ((double)EXTRA_HOPS_AT_BOTTOM + (double) node.maxHTL());
+		double timePerHop = timeout / (EXTRA_HOPS_AT_BOTTOM + node.maxHTL());
 		return (short) Math.min(node.maxHTL(), time / timePerHop);
 	}
